# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: CortexAI POS API
  description: |
    API REST do CortexAI POS — sistema de Ponto de Venda com inteligência artificial.
    Gerencia produtos, vendas, clientes, cupons, categorias, estoque e processamento de texto com IA.
  version: 6.0.0
  contact:
    name: CortexAI
servers:
  - url: http://localhost:3001
    description: Servidor de desenvolvimento
paths:
...
  /api/venda:
    post:
      summary: Registrar nova venda
      tags: [Vendas]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarVenda'
      responses:
        '200':
          description: Venda registrada com sucesso
        '400':
          description: Dados inválidos ou estoque insuficiente
  /api/historico:
    get:
      summary: Histórico de vendas (paginado)
      tags: [Vendas]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: dataInicio
          schema: { type: string, format: date-time }
        - in: query
          name: dataFim
          schema: { type: string, format: date-time }
        - in: query
          name: origem
          schema: { type: string }
        - in: query
          name: produto
          schema: { type: string }
      responses:
        '200':
          description: Histórico de vendas
  /api/vendas:
    get:
      summary: Listar últimas vendas
      tags: [Vendas]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Lista de vendas
  /api/stats:
    get:
      summary: Estatísticas gerais (total vendas, faturamento, ticket médio)
      tags: [Vendas]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Estatísticas
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalVendas: { type: integer }
                  faturamentoTotal: { type: number }
                  ticketMedio: { type: number }
                  produtosUnicos: { type: integer }
                  valorEstoque: { type: number }
  /api/vendas-7dias:
    get:
      summary: Vendas dos últimos 7 dias
      tags: [Vendas]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Vendas por dia

  # ====== CLIENTES ======
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriarVenda:
      type: object
      required: [produto, quantidade, preco]
      properties:
        produto: { type: string }
        quantidade: { type: integer, minimum: 1 }
        preco: { type: number, minimum: 0.01 }
        origem: { type: string, default: manual }
        audioLogId: { type: integer }
        usuarioId: { type: integer }
        clienteId: { type: integer }
        desconto: { type: number, default: 0 }
        metodoPagamento: { type: string }
        troco: { type: number }
        cupomUsado: { type: string }
...